alias evmserverd as manageiq.service
all of our services go by manageiq.
This allows us to start and stop services using the name manageiq

We may need to do something special for the alias to take effect.

Of note, enabling the service using the alias does not work
But I think all other operations do